Refactor: refatora dados do profissional
