+++ frontend/atendimento-app/src/components/shared/header.tsx
@@ -3,31 +3,11 @@
 import Image from "next/image";
 import Link from "next/link";
 import { User } from "lucide-react";
-import { useEffect, useState } from "react";
 
-import { MeusDadosModal } from "../modals/meusDadosModal";
+import { MeusDadosModal } from "../../features/profissional/components/meusDadosModal";
 import { Avatar, AvatarImage, AvatarFallback } from "@/components/ui/avatar";
 
-import { Profissional } from "@/types/Profissional";
-
-import dados from "../../../data/verificacao.json";
-import { getProfissional } from "@/api/dadosProfissional";
-
 export default function Header() {
-  const [profissional, setProfissional] = useState<Profissional | null>(null);
-
-  useEffect(() => {
-    async function carregarProfissional() {
-      try {
-        const data = await getProfissional(dados.idProfissional);
-        setProfissional(data);
-      } catch (error) {
-        console.error("Erro ao buscar profissional", error);
-      }
-    }
-
-    carregarProfissional();
-  }, []);
 
   return (
     <header className="w-full border-b bg-[#F8FAFD] shadow-sm">
@@ -43,24 +23,17 @@ export default function Header() {
           <span className="font-bold text-lg text-[#344054]">APAE</span>
         </Link>
 
-        {profissional && (
-          <MeusDadosModal
-            trigger={
-              <Avatar className="w-10 h-10 rounded-full bg-[#F2F4F7] flex items-center justify-center hover:cursor-pointer">
-                <AvatarImage src="" />
-                <AvatarFallback>
-                  <User className="w-5 h-5 text-[#344054]" />
-                </AvatarFallback>
-              </Avatar>
-            }
-            userData={{
-              nome: profissional.nomeCompleto,
-              crm: profissional.crm,
-              celular: profissional.contato,
-              email: profissional.email,
-            }}
-          />
-        )}
+
+        <MeusDadosModal
+          trigger={
+            <Avatar className="w-10 h-10 rounded-full bg-[#F2F4F7] flex items-center justify-center hover:cursor-pointer">
+              <AvatarImage src="" />
+              <AvatarFallback>
+                <User className="w-5 h-5 text-[#344054]" />
+              </AvatarFallback>
+            </Avatar>
+          }
+        />
       </div>
     </header>
   );
